requirements.txt
faiss==1.7.4
lvis==0.5.3
numpy==1.24.3
Pillow==9.4.0
Pillow==10.1.0
pycocotools==2.0.6
scipy==1.10.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.65.0
typing_extensions==4.9.0
